python/requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.5.0
Werkzeug==2.3.7
Flask-Limiter==3.5.0
pytest==7.4.2
pytest-flask==1.3.0
gunicorn==21.2.0
cryptography==41.0.4 